Fixed RegExp
Added a new Regular Expression for link parsing that should work in Firefox.

The old one used a Lookbehind, which is an experimental RegExp feature :

(?<=@)

Which isn't functional in most browsers.

(and ISN'T supposed to work in Chrome, either. the fact it works is apparently another bug !)

I replaced it with a pattern that just checks if we've got something like this:

(...)/@username/posttitle

and extracts "username" and "posttitle" as before.